Add support for custom delimiters in add function
Introduced functionality to specify custom delimiters at the start of the input string with the format "//[delimiter]\n". Updated the code to dynamically generate a regular expression based on the custom delimiter and split the numbers accordingly. This enhancement allows greater flexibility in the input string format for the add function.